update folders structure
