+++ src/Components/Homepage/MainContentBlock/MainContentTitle.js
@@ -1,0 +1,14 @@
+import dotLine from './images/nuts_red.gif'
+
+function MainContentTitle({title, info, date}) {
+    return (
+        <div className="pt-14">
+            <h2 className="text-4xl text-center text-white">{title}</h2>
+            <img src={dotLine} className="mx-auto my-5"></img>
+            {date && <p className="text-center pb-3 text-white">{date}</p>}
+            <p className="max-w-xl mx-auto mb-5 text-[#ff9fac]">{info}</p>
+        </div>
+    )
+}
+
+export default MainContentTitle;
