+++ src/Components/Homepage/index.js
@@ -1,45 +1,14 @@
-import Header from "./Header";
+import Header from "../Header";
 import MainContentBlock from "./MainContentBlock";
 import InfoBlock from "./InfoBlock";
 import Sponsors from "./Sponsors";
-import Footer from "./Footer";
+import Footer from "../Footer";
 
 const mainContentData = {
     title: "Get Ready to Build",
     date: "NOVEMBER 12-16, 2012 IN BELFAST, NORTHERN IRELAND",
     info: "Build is a festival for people who design for the web. For one week in November, a community of smart, talented people come together to share ideas, tell stories, get their hands dirty, and drink a few beers."
 }
-
-const conferenceSpeakerData = [
-    {
-        profileImg: "./images/robin.png",
-        name: "Robin Sloan"
-    },
-    {
-        profileImg: "./images/kirby.png",
-        name: "Kirby Ferguson"
-    },
-    {
-        profileImg: "./images/rob.png",
-        name: "Rob Giampietro"
-    },
-    {
-        profileImg: "./images/mandy.png",
-        name: "Mandy Brown"
-    },
-    {
-        profileImg: "./images/jeff.png",
-        name: "Jeff Veen"
-    },
-    {
-        profileImg: "./images/tiffani.png",
-        name: "Tiffani Jones Brown"
-    },
-    {
-        profileImg: "./images/ethan.png",
-        name: "Ethan Marcotte"
-    }
-]
 
 const workshopData = {
     title: "Workshop",
@@ -147,7 +116,7 @@ function Homepage() {
     return (
         <>
             <Header />
-            <MainContentBlock title={mainContentData.title} info={mainContentData.info} date={mainContentData.date} conferenceSpeakers={conferenceSpeakerData} />
+            <MainContentBlock title={mainContentData.title} info={mainContentData.info} date={mainContentData.date} />
             <InfoBlock title={workshopData.title} info={workshopData.info} features={workshopData.features} />
             <InfoBlock title={festivalData.title} info={festivalData.info} features={festivalData.features} />
             <InfoBlock title={fringeData.title} info={fringeData.info} features={fringeData.features} />
